+++ plugins/bed/src/util.ts
@@ -1,139 +1,18 @@
 import BED from '@gmod/bed'
-import { SimpleFeature } from '@jbrowse/core/util'
+import { SimpleFeatureSerialized } from '@jbrowse/core/util'
+import {
+  generateBedMethylFeature,
+  isBedMethylFeature,
+} from './generateBedMethylFeature'
+import {
+  generateUcscTranscript,
+  isUcscTranscript,
+} from './generateUcscTranscript'
+import {
+  generateRepeatMaskerFeature,
+  isRepeatMaskerDescriptionField,
+} from './generateRepeatMaskerFeature'
 
-export interface MinimalFeature {
-  type: string
-  start: number
-  end: number
-  refName: string
-  [key: string]: unknown
-}
-
-export interface TranscriptFeat extends MinimalFeature {
-  thickStart: number
-  thickEnd: number
-  blockCount: number
-  blockSizes: number[]
-  chromStarts: number[]
-  refName: string
-  strand?: number
-  subfeatures: MinimalFeature[]
-}
-
-export function generateUcscTranscript(feature: TranscriptFeat) {
-  const {
-    subfeatures: oldSubfeatures,
-    thickStart,
-    thickEnd,
-    blockCount,
-    blockSizes,
-    chromStarts,
-    refName,
-    strand = 0,
-    ...rest
-  } = feature
-
-  if (!thickStart || !thickEnd || !strand) {
-    return feature
-  }
-
-  const subfeatures: MinimalFeature[] = []
-  const feats = oldSubfeatures
-    .filter(child => child.type === 'block')
-    .sort((a, b) => a.start - b.start)
-
-  for (const block of feats) {
-    const start = block.start
-    const end = block.end
-    if (thickStart >= end) {
-      // left-side UTR
-      subfeatures.push({
-        type: `${strand > 0 ? 'five' : 'three'}_prime_UTR`,
-        start,
-        end,
-        refName,
-      })
-    } else if (thickStart > start && thickStart < end && thickEnd >= end) {
-      // UTR | CDS
-      subfeatures.push(
-        {
-          type: `${strand > 0 ? 'five' : 'three'}_prime_UTR`,
-          start,
-          end: thickStart,
-          refName,
-        },
-        {
-          type: 'CDS',
-          start: thickStart,
-          end,
-          refName,
-        },
-      )
-    } else if (thickStart <= start && thickEnd >= end) {
-      // CDS
-      subfeatures.push({
-        type: 'CDS',
-        start,
-        end,
-        refName,
-      })
-    } else if (thickStart > start && thickStart < end && thickEnd < end) {
-      // UTR | CDS | UTR
-      subfeatures.push(
-        {
-          type: `${strand > 0 ? 'five' : 'three'}_prime_UTR`,
-          start,
-          end: thickStart,
-          refName,
-        },
-        {
-          type: 'CDS',
-          start: thickStart,
-          end: thickEnd,
-          refName,
-        },
-        {
-          type: `${strand > 0 ? 'three' : 'five'}_prime_UTR`,
-          start: thickEnd,
-          end,
-          refName,
-        },
-      )
-    } else if (thickStart <= start && thickEnd > start && thickEnd < end) {
-      // CDS | UTR
-      subfeatures.push(
-        {
-          type: 'CDS',
-          start,
-          end: thickEnd,
-          refName,
-        },
-        {
-          type: `${strand > 0 ? 'three' : 'five'}_prime_UTR`,
-          start: thickEnd,
-          end,
-          refName,
-        },
-      )
-    } else if (thickEnd <= start) {
-      // right-side UTR
-      subfeatures.push({
-        type: `${strand > 0 ? 'three' : 'five'}_prime_UTR`,
-        start,
-        end,
-        refName,
-      })
-    }
-  }
-
-  return {
-    ...rest,
-    strand,
-    type: 'mRNA',
-    refName,
-    subfeatures,
-  }
-}
 function stringToStrand(f: string) {
   if (f === '-1') {
     return -1
@@ -208,75 +87,61 @@ export function makeBlocks({
   return []
 }
 
-function generateBedMethylFeature({
+export function featureData({
   line,
+  colRef,
+  colStart,
+  colEnd,
+  scoreColumn,
+  parser,
   uniqueId,
-  refName,
-  start,
-  end,
+  names,
 }: {
   line: string
+  colRef: number
+  colStart: number
+  colEnd: number
+  scoreColumn: string
+  parser: BED
   uniqueId: string
-  refName: string
-  start: number
-  end: number
+  names?: string[]
 }) {
-  // see
-  // https://github.com/nanoporetech/modkit?tab=readme-ov-file#description-of-bedmethyl-output
-  const [
-    ,
-    ,
-    ,
-    code,
-    ,
-    strand,
-    ,
-    ,
-    color,
-    n_valid_cov,
-    fraction_modified,
-    n_mod,
-    n_canonical,
-    n_other_mod,
-    n_delete,
-    n_fail,
-    n_diff,
-    n_nocall,
-  ] = line.split('\t')
-  return new SimpleFeature({
-    id: uniqueId,
-    data: {
-      refName,
-      start,
-      end,
-      code,
-      score: fraction_modified,
-      strand,
-      color,
-      source: code,
-      n_valid_cov,
-      fraction_modified,
-      n_mod,
-      n_canonical,
-      n_other_mod,
-      n_delete,
-      n_fail,
-      n_diff,
-      n_nocall,
-    },
+  const splitLine = line.split('\t')
+  const refName = splitLine[colRef]!
+  const start = +splitLine[colStart]!
+  const end = +splitLine[colEnd]! + (colStart === colEnd ? 1 : 0)
+
+  return featureData2({
+    line,
+    refName,
+    start,
+    end,
+    parser,
+    uniqueId,
+    scoreColumn,
+    names,
   })
 }
 
-export function featureData(
-  line: string,
-  colRef: number,
-  colStart: number,
-  colEnd: number,
-  scoreColumn: string,
-  parser: BED,
-  uniqueId: string,
-  names?: string[],
-) {
+export function featureData2({
+  line,
+  refName,
+  start,
+  end,
+  parser,
+  uniqueId,
+  scoreColumn,
+  names,
+}: {
+  line: string
+  refName: string
+  start: number
+  end: number
+  parser: BED
+  uniqueId: string
+  scoreColumn: string
+  names?: string[]
+}): SimpleFeatureSerialized {
   const splitLine = line.split('\t')
   const data = names
     ? defaultParser(names, line)
@@ -301,12 +166,9 @@ export function featureData(
     description,
     ...rest2
   } = rest
-
-  const refName = splitLine[colRef]!
-  const start = +splitLine[colStart]!
-  const end = +splitLine[colEnd]! + (colStart === colEnd ? 1 : 0)
   const score = scoreColumn ? +data[scoreColumn] : score2 ? +score2 : undefined
   const strand = typeof strand2 === 'string' ? stringToStrand(strand2) : strand2
+
   const subfeatures = makeBlocks({
     start,
     uniqueId,
@@ -326,105 +188,50 @@ export function featureData(
       end,
     })
   } else if (isRepeatMaskerDescriptionField(description)) {
-    return new SimpleFeature({
-      id: uniqueId,
-      data: {
-        ...rest2,
-        ...makeRepeatTrackDescription(description),
-        refName,
-        start,
-        end,
-        type,
-        subfeatures,
-      },
-    })
-  } else if (
-    isUcscTranscript({
-      strand,
-      blockCount,
-      thickStart,
+    return generateRepeatMaskerFeature({
+      ...rest2,
+      uniqueId,
       description,
+      type,
+      score,
+      start,
+      end,
+      strand,
+      refName,
+      subfeatures,
     })
-  ) {
-    const {
-      strand: strand2,
-      score: score2,
-      chrom: _1,
-      chromStart: _2,
-      chromEnd: _3,
+  } else if (isUcscTranscript({ strand, blockCount, thickStart })) {
+    return generateUcscTranscript({
+      ...rest,
+      description,
       chromStarts,
-      blockStarts,
+      thickStart,
+      thickEnd,
       blockSizes,
-      ...rest
-    } = data
-    return new SimpleFeature({
-      id: uniqueId,
-      data: generateUcscTranscript({
-        ...rest,
-        ...makeRepeatTrackDescription(description),
-        chromStarts,
-        thickStart,
-        thickEnd,
-        blockSizes,
-        blockCount,
-        type,
-        score,
-        start,
-        end,
-        strand,
-        refName,
-        uniqueId,
-        subfeatures,
-      }),
+      blockCount,
+      type,
+      score,
+      start,
+      end,
+      strand,
+      refName,
+      uniqueId,
+      subfeatures,
     })
   } else {
-    return new SimpleFeature({
-      id: uniqueId,
-      data: {
-        ...rest,
-        ...makeRepeatTrackDescription(description),
-        type,
-        score,
-        start,
-        end,
-        strand,
-        refName,
-        uniqueId,
-        subfeatures,
-      },
-    })
+    return {
+      ...rest,
+      uniqueId,
+      description,
+      type,
+      score,
+      start,
+      end,
+      strand,
+      refName,
+      subfeatures,
+    }
   }
-}
-
-export function isUcscTranscript({
-  thickStart,
-  blockCount,
-  strand,
-  description,
-}: {
-  thickStart?: number
-  blockCount?: number
-  strand?: number
-  description?: string
-}) {
-  return (
-    thickStart &&
-    blockCount &&
-    strand !== 0 &&
-    !isRepeatMaskerDescriptionField(description)
-  )
-}
-
-export function isBedMethylFeature({
-  splitLine,
-  start,
-  end,
-}: {
-  splitLine: string[]
-  start: number
-  end: number
-}) {
-  return +(splitLine[6] || 0) === start && +(splitLine[7] || 0) === end
 }
 
 export function arrayify(f?: string | number[]) {
@@ -433,50 +240,4 @@ export function arrayify(f?: string | number[]) {
       ? f.split(',').map(f => +f)
       : f
     : undefined
-}
-
-function isRepeatMaskerDescriptionField(desc?: string): desc is string {
-  const ret = desc?.trim().split(' ')
-  return [0, 1, 2, 3, 5, 6].every(s =>
-    ret?.[s] !== undefined ? !Number.isNaN(+ret[s]) : false,
-  )
-}
-export function makeRepeatTrackDescription(description?: string) {
-  if (isRepeatMaskerDescriptionField(description)) {
-    const [
-      bitsw_score,
-      percent_div,
-      percent_del,
-      percent_ins,
-      query_chr,
-      query_begin,
-      query_end,
-      query_remaining,
-      orientation,
-      matching_repeat_name,
-      matching_repeat_class,
-      matching_repeat_begin,
-      matching_repeat_end,
-      matching_repeat_remaining,
-      repeat_id,
-    ] = description.trim().split(' ')
-    return {
-      bitsw_score,
-      percent_div,
-      percent_del,
-      percent_ins,
-      query_chr,
-      query_begin,
-      query_end,
-      query_remaining,
-      orientation,
-      matching_repeat_name,
-      matching_repeat_class,
-      matching_repeat_begin,
-      matching_repeat_end,
-      matching_repeat_remaining,
-      repeat_id,
-    }
-  }
-  return { description }
 }
